Add separate platform for SQL Server 2017
And moved getAggregateConcatExpression there.
